refactor billing settings to use an api call
